+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add Android L Preview Easter Egg [#714](https://github.com/hushenghao/AndroidEasterEggs/issues/714)
 - Add Platlogo snapshot preview
+- Timeline add Android logo
 - Optimization Cat Editor options panel
 - Fix Timeline list scroll gesture conflicts [#654](https://github.com/hushenghao/AndroidEasterEggs/issues/654)
 - Upgrade the minimum supported version to Android 6.0